use same table for both role view and users view
