+++ client/src/components/settings/BackButtons.tsx
@@ -48,11 +48,7 @@ export function SidebarBack() {
       variant="ghost"
       className="justify-start"
       onClick={() => {
-        if (router.history.length > 0) {
-          router.history.back();
-        } else {
-          router.navigate({ to: "/chat" });
-        }
+        router.navigate({ to: "/chat" });
       }}
     >
       <ArrowLeft />
