+++ client/src/routes/admin/users/$userId.tsx
@@ -1,9 +1,17 @@
-import { createFileRoute } from '@tanstack/react-router'
+import { PageBack } from "@/components/settings/BackButtons";
+import { createFileRoute } from "@tanstack/react-router";
 
-export const Route = createFileRoute('/admin/users/$userId')({
+export const Route = createFileRoute("/admin/users/$userId")({
   component: RouteComponent,
-})
+});
 
 function RouteComponent() {
-  return <div>Hello "/admin/users/$userId"!</div>
+  const { userId } = Route.useParams();
+
+  return (
+    <div className="flex w-full flex-col gap-2 p-4">
+      <PageBack />
+      <h1 className="p-2 text-2xl font-bold">Edit Role - {role.data?.name ?? "Loading..."}</h1>
+    </div>
+  );
 }
